requirements.txt
openpyxl==3.1.5
Pillow==12.1.1
